% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_twetes.R
\name{get_twetes}
\alias{get_twetes}
\title{Helper function to grab data from the Twitter API}
\usage{
get_twetes(endpoint, token_header, api_query, lst = NULL)
}
\arguments{
\item{endpoint}{API endpoint to use}

\item{token_header}{Authentication header to pass to httr}

\item{api_query}{query to send to the API}

\item{lst}{default NULL. Data accumulated so far, to be kept}
}
\value{
list of tweet data
}
\description{
Parameters are set by the enclosing parent function \code{get_user_twetes}
}
