% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_id.R
\name{get_user_id}
\alias{get_user_id}
\title{Get a twitter user ID for a twitter username}
\usage{
get_user_id(username)
}
\arguments{
\item{username}{Twitter username as a string (no \code{@} required)}
}
\value{
the user ID as a string (check! might be as a numeric!)
}
\description{
Use the Twitter API to find the User ID for a given Twitter username
}
