% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_twetes.R
\name{get_user_twetes}
\alias{get_user_twetes}
\title{Setup script to retrieve tweets}
\usage{
get_user_twetes(user_id, start)
}
\arguments{
\item{user_id}{Twitter user ID as a string}

\item{start}{Start date - from when to retrieve tweet data. Needs to be in a particular ISO8601 format like \code{YYYY-MM-DDTHH:mm:ssZ} (check this!)}
}
\value{
list of tweet data
}
\description{
This is a parent script to \code{get_twetes.R}. Here we set the desired fields to return, set up the specific endpoint for the user id we are interested in, deal with authentication and create a query to send to the API.
}
