% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{unpack_list}
\alias{unpack_list}
\title{Tidying helper function: rectangularises a nested list}
\usage{
unpack_list(lst)
}
\arguments{
\item{lst}{a list (presumably nested)}
}
\value{
a tibble
}
\description{
Turns the list data from the Twitter API into a rectangular tibble
}
