% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_twetes.R
\name{get_user_twetes}
\alias{get_user_twetes}
\title{Setup script to retrieve tweets}
\usage{
get_user_twetes(user_id, start)
}
\arguments{
\item{user_id}{Twitter user ID as a string}

\item{start}{Start date - from when to retrieve tweet data. Needs to be in a particular ISO8601 format like \code{YYYY-MM-DDTHH:mm:ssZ} (check this!)}
}
\value{
list of tweet data
}
\description{
Setup script to retrieve tweets
}
